+++ rapl-configure/rapl-configure.c
@@ -19,6 +19,8 @@ typedef struct rapl_configure_ctx {
   int get_sockets;
   raplcap_zone zone;
   unsigned int socket;
+  int enabled;
+  int set_enabled;
   int set_long;
   double watts_long;
   double sec_long;
@@ -29,11 +31,12 @@ typedef struct rapl_configure_ctx {
 
 static rapl_configure_ctx ctx;
 static const char* prog;
-static const char short_options[] = "nc:z:s:w:S:W:h";
+static const char short_options[] = "nc:z:e:s:w:S:W:h";
 static const struct option long_options[] = {
   {"nsockets", no_argument,       NULL, 'n'},
   {"socket",   required_argument, NULL, 'c'},
   {"zone",     required_argument, NULL, 'z'},
+  {"enabled",  required_argument, NULL, 'e'},
   {"seconds0", required_argument, NULL, 's'},
   {"watts0",   required_argument, NULL, 'w'},
   {"seconds1", required_argument, NULL, 'S'},
@@ -54,13 +57,15 @@ static void print_usage(int exit_code) {
           "                           UNCORE - uncore power plane (client systems only)\n"
           "                           DRAM - main memory (server systems only)\n"
           "                           PSYS - the entire platform (Skylake and newer only)\n"
-          "  -s, --seconds0=SECONDS   long term time window\n"
-          "  -w, --watts0=WATTS       long term power limit\n"
-          "  -S, --seconds1=SECONDS   short term time window (PACKAGE & PSYS only)\n"
-          "  -W, --watts1=WATTS       short term power limit (PACKAGE & PSYS only)\n"
+          "  -e, --enabled=1|0        Enable/disable a zone\n"
+          "  -s, --seconds0=SECONDS   Long term time window\n"
+          "  -w, --watts0=WATTS       Long term power limit\n"
+          "  -S, --seconds1=SECONDS   Short term time window (PACKAGE & PSYS only)\n"
+          "  -W, --watts1=WATTS       Short term power limit (PACKAGE & PSYS only)\n"
           "  -h, --help               Print this message and exit\n\n"
-          "Unless time or power limits are specified, current values will be printed.\n"
-          "If the only values specified are 0, the zone will be disabled.\n\n",
+          "Current values are printed if no flags, or only socket and/or zone flags, are specified.\n"
+          "Otherwise, specified values are set while other values remain unmodified.\n"
+          "When setting values, zones are automatically enabled unless -e/--enabled is explicitly set to 0.\n",
           prog);
   exit(exit_code);
 }
@@ -97,8 +102,8 @@ static void print_limits(int enabled,
   }
 }
 
-static void print_enable_error(const char* fn) {
-  perror(fn);
+static void print_error_continue(const char* msg) {
+  perror(msg);
   fprintf(stderr, "Trying to proceed anyway...\n");
 }
 
@@ -108,27 +113,34 @@ static int configure_limits(const rapl_configure_ctx* c) {
   raplcap_limit limit_short;
   raplcap_limit* ll = NULL;
   raplcap_limit* ls = NULL;
-  int disable = 1;
+  int ret = 0;
+  int disable = ctx.set_enabled && !ctx.enabled;
   if (c->set_long) {
     limit_long.seconds = c->sec_long;
     limit_long.watts = c->watts_long;
     ll = &limit_long;
-    disable &= is_zero_dbl(limit_long.seconds) && is_zero_dbl(limit_long.watts);
   }
   if (c->set_short) {
     limit_short.seconds = c->sec_short;
     limit_short.watts = c->watts_short;
     ls = &limit_short;
-    disable &= is_zero_dbl(limit_short.seconds) && is_zero_dbl(limit_short.watts);
   }
-  if (disable) {
-    // all given values were 0 - disable the zone
-    return raplcap_set_zone_enabled(NULL, c->socket, c->zone, 0);
+  // disable zone first, if requested
+  if (disable && (ret = raplcap_set_zone_enabled(NULL, c->socket, c->zone, 0))) {
+    perror("Failed to disable zone");
+    return ret;
   }
-  if (raplcap_set_zone_enabled(NULL, c->socket, c->zone, 1)) {
-    print_enable_error("raplcap_set_zone_enabled");
+  // set limits
+  if ((c->set_long || c->set_short) && (ret = raplcap_set_limits(NULL, c->socket, c->zone, ll, ls))) {
+    perror("Failed to set limits");
+    return ret;
   }
-  return raplcap_set_limits(NULL, c->socket, c->zone, ll, ls);
+  // enable zone (if not disable)
+  if (!disable && (ret = raplcap_set_zone_enabled(NULL, c->socket, c->zone, 1))) {
+    perror("Failed to enable zone");
+    return ret;
+  }
+  return 0;
 }
 
 static int get_limits(unsigned int socket, raplcap_zone zone) {
@@ -136,21 +148,30 @@ static int get_limits(unsigned int socket, raplcap_zone zone) {
   raplcap_limit ls;
   double joules;
   double joules_max;
+  int ret;
   memset(&ll, 0, sizeof(raplcap_limit));
   memset(&ls, 0, sizeof(raplcap_limit));
   int enabled = raplcap_is_zone_enabled(NULL, socket, zone);
   if (enabled < 0) {
-    print_enable_error("raplcap_is_zone_enabled");
+    print_error_continue("Failed to determine if zone is enabled");
   }
-  if (raplcap_get_limits(NULL, socket, zone, &ll, &ls)) {
-    return -1;
+  if ((ret = raplcap_get_limits(NULL, socket, zone, &ll, &ls))) {
+    perror("Failed to get limits");
+    return ret;
   }
   // we'll consider energy counter information to be optional
   joules = raplcap_get_energy_counter(NULL, socket, zone);
   joules_max = raplcap_get_energy_counter_max(NULL, socket, zone);
   print_limits(enabled, ll.watts, ll.seconds, ls.watts, ls.seconds, joules, joules_max);
-  return 0;
+  return ret;
 }
+
+#define SET_VAL(optarg, val, set_val) \
+  if ((val = atof(optarg)) <= 0) { \
+    fprintf(stderr, "Time window and power limit values must be > 0\n"); \
+    print_usage(1); \
+  } \
+  set_val = 1
 
 int main(int argc, char** argv) {
   int ret = 0;
@@ -188,21 +209,21 @@ int main(int argc, char** argv) {
           print_usage(1);
         }
         break;
+      case 'e':
+        ctx.enabled = atoi(optarg);
+        ctx.set_enabled = 1;
+        break;
       case 's':
-        ctx.sec_long = atof(optarg);
-        ctx.set_long = 1;
+        SET_VAL(optarg, ctx.sec_long, ctx.set_long);
         break;
       case 'w':
-        ctx.watts_long = atof(optarg);
-        ctx.set_long = 1;
+        SET_VAL(optarg, ctx.watts_long, ctx.set_long);
         break;
       case 'S':
-        ctx.sec_short = atof(optarg);
-        ctx.set_short = 1;
+        SET_VAL(optarg, ctx.sec_short, ctx.set_short);
         break;
       case 'W':
-        ctx.watts_short = atof(optarg);
-        ctx.set_short = 1;
+        SET_VAL(optarg, ctx.watts_short, ctx.set_short);
         break;
       case '?':
       default:
@@ -216,21 +237,15 @@ int main(int argc, char** argv) {
   if (ctx.get_sockets) {
     sockets = raplcap_get_num_sockets(NULL);
     if (sockets == 0) {
-      perror("raplcap_get_num_sockets");
+      perror("Failed to get number of sockets");
       return 1;
     }
     printf("%"PRIu32"\n", sockets);
     return 0;
   }
 
-  // verify parameters
-  if (ctx.watts_short < 0 || ctx.sec_short < 0 || ctx.watts_long < 0 || ctx.sec_long < 0) {
-    fprintf(stderr, "Power and interval values must be >= 0\n");
-    print_usage(1);
-  }
-
   // initialize
-  is_read_only = !ctx.set_long && !ctx.set_short;
+  is_read_only = !ctx.set_enabled && !ctx.set_long && !ctx.set_short;
 #ifndef _WIN32
   if (is_read_only) {
     // request read-only access (not supported by all implementations, therefore not guaranteed)
@@ -238,7 +253,7 @@ int main(int argc, char** argv) {
   }
 #endif
   if (raplcap_init(NULL)) {
-    perror("Init failed");
+    perror("Failed to initialize");
     return 1;
   }
 
@@ -248,8 +263,7 @@ int main(int argc, char** argv) {
     ret = -1;
   } else {
     if (supported < 0) {
-      perror("raplcap_is_zone_supported");
-      fprintf(stderr, "Trying to proceed anyway...\n");
+      print_error_continue("Failed to determine if zone is supported");
     }
     // perform requested action
     if (is_read_only) {
@@ -257,14 +271,11 @@ int main(int argc, char** argv) {
     } else {
       ret = configure_limits(&ctx);
     }
-    if (ret) {
-      perror("Action failed");
-    }
   }
 
   // cleanup
   if (raplcap_destroy(NULL)) {
-    perror("Cleanup failed");
+    perror("Failed to clean up");
   }
 
   return ret;