+++ RELEASES.md
@@ -3,11 +3,13 @@
 ## [Unreleased]
 ### Added
  * MSR implementation support for Atom processors
- * Functions for reading energy counters
+ * Functions for reading energy counters (also added output to rapl-configure binaries)
+ * New -e/--enabled flag for rapl-configure binaries
  * More units tests
 
 ### Changed
  * MSR implementation will now only compile on x86 (required for RAPL anyway)
+ * Cannot disable zones by setting time window or power limit values to 0 in rapl-configure binaries (use -e/--enabled=0 instead)
 
 
 ## [v0.2.0] - 2017-12-11